+++ client/src/App.js
@@ -1,10 +1,12 @@
 import React from 'react';
 import './App.scss';
 import { BrowserRouter, Switch, Route, Link, NavLink } from 'react-router-dom';
-import UserList from "./UserList";
+import AccountManager from "./AccountManager";
 import MusicList from "./MusicList";
 import Home from "./Home";
 import InputFormMusic from './InputFormMusic';
+import LogIn from './LogIn';
+import SignUp from './SignUp';
 import Error from "./Error404";
 
 function App() {
@@ -28,8 +30,10 @@ function App() {
       <Switch>
         <Route exact path="/" component={Home} />
         <Route exact path="/music" component={MusicList} />
-        <Route exact path="/user" component={UserList} />
+        <Route exact path="/user" component={AccountManager} />
         <Route exact path="/music/post" component={InputFormMusic} />
+        <Route exact path="/user/login" component={LogIn} />
+        <Route exact path="/user/signup" component={SignUp} />
         <Route component={Error} />
       </Switch>
